Class File API - Add validation of corner cases. Support arbitrary method signatures for proxying.

+++ class-file-api/src/main/java/org/bananalaba/jdk24/CalculatorService.java
@@ -2,7 +2,7 @@ package org.bananalaba.jdk24;
 
 public class CalculatorService {
 
-    public int add(final int a, final int b) {
+    public float add(final float a, final int b) {
         return a + b;
     }
